+++ src/app/models/dataTransfer.ts
@@ -19,5 +19,5 @@ export interface DataTransfer {
   modified: string;
   version: number;
   downloaded?: boolean;
-  downloading?: boolean;
+  downloadInProgress?: boolean;
 }